Add files via upload
Now a days there has been a significant increase in traffic accidents caused by driver fatigue which results in several fatalities. The number of accidents can be reduced by having a proper system that can detect drowsiness, alert the driver and prevent major injuries. There has been an increase in safety systems in cars and other vehicles and many are now mandatory in vehicles , but all of them cannot help if a driver falls asleep for a brief moment. Hence that is what we are going to build today – DRIVER SLEEP DETECTION SYSTEM. This can serve as the foundation for a system that could reduce accidents caused by driving drowsiness. In this project we will use few libraries from python which will help in detecting the sleep if eyes are closed  and also the library will provide alarm sounds after it detects sleep. If the driver eyes are closed then the camera detects it / senses it and then the sound automatically starts after detection, which will help the driver be safe. The techniques used in this project are: Eye Tracking, Facial Recognition, Machine Learning, Alerting Mechanisms.

Libraries: Open CV, Numpy, DLIB.